requirements.txt
Flask
pandas
plotly
pymysql
sqlalchemy